# openapi.yaml
openapi: 3.0.1
info:
  title: '"TodoList API"'
  description: TodoList API challenge
  license:
    name: Apache 2.0
    url: https://springdoc.org
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /items:
    get:
      tags:
      - items-endpoint
      summary: Returns a list of items filtered by status
      operationId: find
      parameters:
      - name: "false"
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemResponse'
    post:
      tags:
      - items-endpoint
      summary: Create a new item
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
        required: true
      responses:
        "400":
          description: Invalid request body supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ItemResponse'
  /items/{id}/update-description:
    patch:
      tags:
      - items-endpoint
      summary: Update item description by id
      operationId: updateItemDescription
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemDescriptionRequest'
        required: true
      responses:
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "400":
          description: Invalid request body supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "200":
          description: OK
  /items/{id}/re-open-item:
    patch:
      tags:
      - items-endpoint
      summary: Update the item status to NOT_DONE by id
      operationId: reOpen
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "200":
          description: OK
  /items/{id}/complete-item:
    patch:
      tags:
      - items-endpoint
      summary: Update the item status to DONE by id
      operationId: complete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "200":
          description: OK
  /items/{id}:
    get:
      tags:
      - items-endpoint
      summary: Return item details by id
      operationId: details
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ItemResponse'
components:
  schemas:
    CreateItemRequest:
      required:
      - description
      - dueAt
      type: object
      properties:
        description:
          type: string
        dueAt:
          type: string
          format: date-time
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
    ItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        dueAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - NOT_DONE
          - DONE
          - PAST_DUE
    UpdateItemDescriptionRequest:
      required:
      - description
      type: object
      properties:
        description:
          type: string